Changed route name to match new capability
